requirements.txt
django==1.9.6
python-dateutil==2.5.3